tests pass on abstract feauture
